Move theme.ts
