+++ src/pages/_document.tsx
@@ -2,7 +2,7 @@
 
 import { ColorModeScript } from "@chakra-ui/react";
 import { Head, Html, Main, NextScript } from "next/document";
-import theme from "./theme";
+import theme from "../theme";
 
 export default function Document() {
   return (
